+++ src/core/pixel.cpp
@@ -15,10 +15,9 @@ namespace PixelMaestro {
 		current_color_ = {0, 0, 0};
 
 #ifndef DISABLE_COLOR_BUFFER
-		step_.r = 0;
-		step_.g = 0;
-		step_.b = 0;
-		step_.count = 0;
+		step_[0] = 0;
+		step_[1] = 0;
+		step_[2] = 0;
 #endif
 	}
 
@@ -39,10 +38,9 @@ namespace PixelMaestro {
 	*/
 	void Pixel::set_next_color(const Colors::RGB& next_color, uint8_t step_count) {
 #ifndef DISABLE_COLOR_BUFFER
-		step_.r = (next_color.r - current_color_.r) / (float)step_count;
-		step_.g = (next_color.g - current_color_.g) / (float)step_count;
-		step_.b = (next_color.b - current_color_.b) / (float)step_count;
-		step_.count = step_count;
+		step_[0] = (next_color.r - current_color_.r) / (float)step_count;
+		step_[1] = (next_color.g - current_color_.g) / (float)step_count;
+		step_[2] = (next_color.b - current_color_.b) / (float)step_count;
 #else
 		current_color_ = next_color;
 #endif
@@ -55,13 +53,9 @@ namespace PixelMaestro {
 	void Pixel::update() {
 #ifndef DISABLE_COLOR_BUFFER
 		// WARNING: This can be imprecise, especially with small or gradual color changes.
-		if (step_.count > 0) {
-			current_color_.r += step_.r;
-			current_color_.g += step_.g;
-			current_color_.b += step_.b;
-
-			step_.count--;
-		}
+		current_color_.r += step_[0];
+		current_color_.g += step_[1];
+		current_color_.b += step_[2];
 #endif
 	}
 }